Migrate to idelista packages
