--- .travis.yml
@@ -1,1 +1,0 @@
-language: java
