+++ build.gradle
@@ -1,11 +1,9 @@
-group 'me.nibbleapp'
-version '0.15'
+group 'com.idealista.android'
+version '1.0-SNAPSHOT'
 
 buildscript {
     ext.kotlin_version = '1.1.51'
-    ext.coroutines_version = '0.17'
     ext.junit_version = '4.12'
-    ext.mockito_version = '2.8.9'
 
     repositories {
         mavenCentral()
@@ -15,30 +13,20 @@ buildscript {
     }
 }
 
-apply plugin: 'java'
 apply plugin: 'kotlin'
-
-sourceCompatibility = 1.7
-
-kotlin {
-    experimental {
-        coroutines 'enable'
-    }
-}
 
 repositories {
     mavenCentral()
-    jcenter()
-}
-
-sourceSets {
-    main.java.srcDirs += 'src/main/kotlin/'
-    test.java.srcDirs += 'src/test/kotlin/'
 }
 
 dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
-    compile 'org.jetbrains.kotlinx:kotlinx-coroutines-core:0.17'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    testCompile "org.mockito:mockito-core:2.8.9"
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
+    testCompile "junit:junit:$junit_version"
+}
+
+compileKotlin {
+    kotlinOptions.jvmTarget = "1.8"
+}
+compileTestKotlin {
+    kotlinOptions.jvmTarget = "1.8"
 }
