--- sample/build.gradle
@@ -1,27 +1,0 @@
-group 'me.nibbleapp.wrench'
-version '1.0-SNAPSHOT'
-
-buildscript {
-    repositories {
-        jcenter()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    }
-}
-
-apply plugin: 'org.jetbrains.kotlin.jvm'
-
-repositories {
-    jcenter()
-}
-
-dependencies {
-
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
-    compile "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinx_coroutines_core_version"
-}
-
-sourceSets {
-    main.java.srcDirs += 'src/main/kotlin/'
-}
