--- sample/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Wed Jul 19 23:59:34 CEST 2017
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-3.3-all.zip
