--- sample/src/main/kotlin/me/nibbleapp/wrench/sample/error/SendEmailError.kt
@@ -1,9 +1,0 @@
-package me.nibbleapp.wrench.sample.error
-
-sealed class SendEmailError {
-
-    class NoNetworkError : SendEmailError()
-    class ServerError : SendEmailError()
-    class BadRequest : SendEmailError()
-
-}
