--- sample/src/main/kotlin/me/nibbleapp/wrench/sample/model/Message.kt
@@ -1,3 +1,0 @@
-package me.nibbleapp.wrench.sample.model
-
-data class Message(val text: String, val date: Long)
