+++ src/main/kotlin/com/idealista/android/wrench/logger/LogFormatter.kt
@@ -1,0 +1,9 @@
+package com.idealista.android.wrench.logger
+
+interface LogFormatter {
+
+  val DEFAULT_STACK_LENGTH: Int
+
+  fun format(message: String, throwable: Throwable?, depth: Int = DEFAULT_STACK_LENGTH): String
+
+}
