+++ src/main/kotlin/com/idealista/android/wrench/usecase/UseCaseExecutable.kt
@@ -1,0 +1,14 @@
+package com.idealista.android.wrench.usecase
+
+import com.idealista.android.wrench.executor.UseCaseExecutor
+import com.idealista.android.wrench.type.Either
+
+class UseCaseExecutable<Error, Result> internal constructor(private val function: () -> Either<Error, Result>,
+                                                            private val ui: (Either<Error, Result>) -> Unit,
+                                                            private val delay: Long) {
+
+    fun run(useCaseExecutor: UseCaseExecutor) =
+            useCaseExecutor.execute(function, ui, delay)
+
+
+}
