--- src/main/kotlin/me/nibbleapp/wrench/logger/Priority.kt
@@ -1,12 +1,0 @@
-package me.nibbleapp.wrench.logger
-
-sealed class Priority(val value: Int) {
-
-  class VERBOSE : Priority(2)
-  class DEBUG : Priority(3)
-  class INFO : Priority(4)
-  class WARN : Priority(5)
-  class ERROR : Priority(6)
-  class ASSERT : Priority(7)
-
-}
