--- src/main/kotlin/me/nibbleapp/wrench/usecase/UseCase.kt
@@ -1,11 +1,0 @@
-package me.nibbleapp.wrench.usecase
-
-import me.nibbleapp.wrench.type.Either
-
-class UseCase<Error, Result> {
-
-    fun bg(background: () -> Either<Error, Result>, delay: Long = 0): Then<Error, Result> =
-            Then(background, delay)
-
-}
-
