+++ src/test/kotlin/com/idealista/android/wrench/type/EitherTest.kt
@@ -1,0 +1,5 @@
+package com.idealista.android.wrench.type
+
+import org.junit.Assert.*
+
+class EitherTest
